Matrix: made dimension and number of nonzeros size_t (unsigned integer)

+++ uno/linear_algebra/SymmetricMatrix.cpp
@@ -5,7 +5,7 @@
 #include "SymmetricMatrix.hpp"
 #include "Vector.hpp"
 
-SymmetricMatrix::SymmetricMatrix(int dimension, size_t capacity) : dimension(dimension), capacity(capacity) {
+SymmetricMatrix::SymmetricMatrix(size_t dimension, size_t capacity) : dimension(dimension), capacity(capacity) {
 }
 
 void SymmetricMatrix::reset() {
